+++ CotEditor/Sources/CEDocumentAnalyzer.m
@@ -35,6 +35,7 @@
 
 #import "NSString+CECounting.h"
 #import "NSString+CEEncoding.h"
+#import "NSString+CENewLine.h"
 
 
 // notifications
@@ -166,15 +167,16 @@ NSString *_Nonnull const CEAnalyzerDidUpdateEditorInfoNotification = @"CEAnalyze
     BOOL needsAll = [self needsUpdateEditorInfo];
     
     CEDocument *document = [self document];
-    CEEditorWrapper *editor = [document editor];
+    NSTextView *textView = [[document editor] focusedTextView];
     
-    BOOL hasMarked = [[editor focusedTextView] hasMarkedText];
-    NSString *wholeString = [document string];
-    NSString *selectedString = hasMarked ? nil : [editor substringWithSelection];
-    __block NSRange selectedRange = [editor selectedRange];
+    if (![textView string]) { return; }
+    
+    NSString *wholeString = [NSString stringWithString:[textView string]];  // LF
+    NSRange selectedRange = [textView selectedRange];
+    CENewLineType lineEnding = [document lineEnding];
     
     // IM で変換途中の文字列は選択範囲としてカウントしない (2007-05-20)
-    if (hasMarked) {
+    if ([textView hasMarkedText]) {
         selectedRange.length = 0;
     }
     
@@ -186,14 +188,17 @@ NSString *_Nonnull const CEAnalyzerDidUpdateEditorInfoNotification = @"CEAnalyze
         
         NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
         BOOL countsLineEnding = [defaults boolForKey:CEDefaultCountLineEndingAsCharKey];
-        NSUInteger column = 0, currentLine = 0, length = [wholeString length], location = 0;
+        NSUInteger column = 0, currentLine = 0, location = 0;
+        NSUInteger length = 0, selectedLength = 0;
         NSUInteger numberOfLines = 0, numberOfSelectedLines = 0;
         NSUInteger numberOfChars = 0, numberOfSelectedChars = 0;
         NSUInteger numberOfWords = 0, numberOfSelectedWords = 0;
         NSString *unicode;
         
-        if (length > 0) {
+        if (wholeString.length > 0) {
+            NSString *selectedString = [wholeString substringWithRange:selectedRange];  // LF
             BOOL hasSelection = (selectedRange.length > 0);
+            
             NSRange lineRange = [wholeString lineRangeForRange:selectedRange];
             column = selectedRange.location - lineRange.location;  // as length
             column = [[wholeString substringWithRange:NSMakeRange(lineRange.location, column)] numberOfComposedCharacters];
@@ -237,10 +242,16 @@ NSString *_Nonnull const CEAnalyzerDidUpdateEditorInfoNotification = @"CEAnalyze
                 }
             }
             
-            // re-calculate length without line ending if needed
-            if (!countsLineEnding) {
-                length = [[wholeString stringByDeletingNewLineCharacters] length];
-                selectedRange.length = [[selectedString stringByDeletingNewLineCharacters] length];
+            // count length
+            if (needsAll || [defaults boolForKey:CEDefaultShowStatusBarLengthKey]) {
+                BOOL isSingleLineEnding = ([[NSString newLineStringWithType:lineEnding] length] == 1);
+                NSString *str = isSingleLineEnding ? wholeString : [wholeString stringByReplacingNewLineCharacersWith:lineEnding];
+                length = [str length];
+                
+                if (hasSelection) {
+                    NSString *str = isSingleLineEnding ? selectedString : [selectedString stringByReplacingNewLineCharacersWith:lineEnding];
+                    selectedLength = [str length];
+                }
             }
             
             if (needsAll) {
@@ -255,7 +266,7 @@ NSString *_Nonnull const CEAnalyzerDidUpdateEditorInfoNotification = @"CEAnalyze
         // apply to UI
         dispatch_sync(dispatch_get_main_queue(), ^{
             self.lines = [self formatCount:numberOfLines selected:numberOfSelectedLines];
-            self.length = [self formatCount:length selected:selectedRange.length];
+            self.length = [self formatCount:length selected:selectedLength];
             self.chars = [self formatCount:numberOfChars selected:numberOfSelectedChars];
             self.words = [self formatCount:numberOfWords selected:numberOfSelectedWords];
             self.location = [NSString localizedStringWithFormat:@"%li", location];